[Android.Runtime] Use JniTypeManager to register natives
The new code uses `JniTypeManager`'s `RegisterNativeMembers` method to
register native marshaling methods. The type manager itself uses the
register method generated by *jnimarshalmethod-gen* tool or method(s)
of the given type, which have
`JniAddNativeMethodRegistrationAttribute`.

When no such register method exists (`registered` = *false*), we
fallback to the original (old) code.

Example register method generated by the *jnimarshalmethod-gen* tool:

```
[JniAddNativeMethodRegistration]
public static void __RegisterNativeMembers (JniNativeMethodRegistrationArguments args)
{
	args.AddRegistrations (new JniNativeMethodRegistration[] {
		new JniNativeMethodRegistration ("n_onCreate", "(Landroid/os/Bundle;)V", new Action<IntPtr, IntPtr, IntPtr> (MainActivity.__<$>_jni_marshal_methods.n_onCreate_Landroid_os_Bundle_))
	});
}
```
